requirements.txt
# NotebookLM Python Microservice Dependencies

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# NotebookLM Integration
notebooklm-py[browser]>=0.3.0

# HTTP Client
httpx>=0.25.0
aiohttp>=3.9.0

# Database (for user-notebook mapping)
motor>=3.3.0
pymongo>=4.6.0

# Utilities
python-dotenv>=1.0.0
structlog>=23.2.0

# Language Detection
lingua-language-detector>=2.1.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
ruff>=0.1.0
